Separate crates for common database definitions, PostgreSQL and MySQL specific definitions - to run both databases simultaneously
