+++ Cargo.toml
@@ -3,7 +3,6 @@ members = [
     # Utils
     "src/utils/container-runtime",
     "src/utils/data-utils",
-    "src/utils/database-gateway",
     "src/utils/datafusion-cli",
     "src/utils/enum-variants",
     "src/utils/event-bus",
@@ -13,6 +12,10 @@ members = [
     "src/utils/multiformats",
     "src/utils/repo-tools",
     "src/utils/tracing-perfetto",
+    # Database
+    "src/database/database-common",
+    "src/database/sqlx-postgres",
+    "src/database/sqlx-mysql",
     # Domain
     "src/domain/opendatafabric",
     "src/domain/core",
@@ -39,7 +42,6 @@ resolver = "2"
 
 # Utils
 container-runtime = { version = "0.165.0", path = "src/utils/container-runtime", default-features = false }
-database-gateway = { version = "0.165.0", path="src/utils/database-gateway", default-features = false }
 enum-variants = { version = "0.165.0", path = "src/utils/enum-variants", default-features = false }
 event-bus = { version = "0.165.0", path = "src/utils/event-bus", default-features = false }
 event-sourcing = { version = "0.165.0", path = "src/utils/event-sourcing", default-features = false }
@@ -49,6 +51,10 @@ multiformats = { version = "0.165.0", path = "src/utils/multiformats", default-f
 kamu-data-utils = { version = "0.165.0", path = "src/utils/data-utils", default-features = false }
 kamu-datafusion-cli = { version = "0.165.0", path = "src/utils/datafusion-cli", default-features = false }
 tracing-perfetto = { version = "0.165.0", path = "src/utils/tracing-perfetto", default-features = false }
+# Database
+database-common = { version = "0.165.0", path="src/database/database-common", default-features = false }
+database-sqlx-postgres = { version = "0.165.0", path="src/database/sqlx-postgres", default-features = false }
+database-sqlx-mysql = { version = "0.165.0", path="src/database/sqlx-mysql", default-features = false }
 # Domain
 opendatafabric = { version = "0.165.0", path = "src/domain/opendatafabric", default-features = false }
 kamu-core = { version = "0.165.0", path = "src/domain/core", default-features = false }
