+++ src/database/database-common/src/accounts_model.rs
@@ -7,25 +7,28 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use thiserror::Error;
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-/*#[async_trait::async_trait]
-pub trait DatabaseTransaction {
-    /// Commits transaction
-    async fn commit(mut self) -> Result<(), DatabaseTransactionError>;
+#[derive(Debug, Serialize, Deserialize, sqlx::Type)]
+#[sqlx(type_name = "account_origin", rename_all = "lowercase")]
+pub enum AccountOrigin {
+    Cli,
+    Github,
+}
 
-    /// Rollbacks transaction
-    async fn rollback(mut self) -> Result<(), DatabaseTransactionError>;
-}*/
-
-/////////////////////////////////////////////////////////////////////////////////////////
-
-#[derive(Error, Debug)]
-pub enum DatabaseTransactionError {
-    #[error(transparent)]
-    SqlxError(sqlx::Error),
+#[derive(Debug, sqlx::FromRow)]
+#[allow(dead_code)]
+pub struct AccountModel {
+    pub id: Uuid,
+    pub email: String,
+    pub account_name: String,
+    pub display_name: String,
+    pub origin: AccountOrigin,
+    pub registered_at: DateTime<Utc>,
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////
