+++ src/database/sqlx-postgres/src/lib.rs
@@ -7,6 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod mysql_db_connection_pool;
+mod postgres_connection_pool;
+mod postgres_dummy_test;
 
-pub use mysql_db_connection_pool::*;
+pub use postgres_connection_pool::*;
+pub use postgres_dummy_test::*;
